Backend/requirements.txt
fastapi
uvicorn[standard]
spotipy
neo4j
redis
anthropic
pydantic
python-dotenv
requests
numpy
certifi
